圣遗物基础

+++ src/main/java/com/yuo/PaiMeng/Capability/IBlowCapability.java
@@ -7,15 +7,15 @@ import net.minecraftforge.common.util.INBTSerializable;
 public interface IBlowCapability extends INBTSerializable<CompoundNBT> {
     /**
      * 设置暴击率
-     * @param criticalRate 0 ~ 1.0f
+     * @param criticalRate 0 ~ 1.0d
      */
-    void setCriticalRate(float criticalRate);
+    void setCriticalRate(double criticalRate);
 
     /**
      * 设置暴击伤害
      * @param criticalDamage 0 ~ infinity
      */
-    void setCriticalDamage(float criticalDamage);
+    void setCriticalDamage(double criticalDamage);
 
     /**
      * 重置双爆属性为默认值
@@ -26,11 +26,11 @@ public interface IBlowCapability extends INBTSerializable<CompoundNBT> {
      * 获取当前暴击率
      * @return 暴击率值
      */
-    float getCriticalRate();
+    double getCriticalRate();
 
     /**
      * 获取当前暴击伤害
      * @return 暴击伤害值
      */
-    float getCriticalDamage();
+    double getCriticalDamage();
 }
